chore: clean up lib further
