+++ src/lib/index.ts
@@ -1,7 +1,9 @@
 export * from './analysis'
-export * from './stockfish'
-export * from './customAnalysis'
-export * from './train/utils'
-export * from './ratingUtils'
+export * from './analytics'
 export * from './common'
+export * from './favorites'
 export * from './lichess'
+export * from './puzzle'
+export * from './ratingUtils'
+export * from './settings'
+export * from './sound'
