+++ src/lib/sound.ts
@@ -1,10 +1,3 @@
-/**
- * Global Chess Sound Manager
- *
- * Centralized audio management for chess moves with efficient resource pooling
- * and mobile-optimized performance.
- */
-
 type SoundType = 'move' | 'capture'
 
 interface AudioPool {
@@ -14,7 +7,7 @@ interface AudioPool {
 
 class ChessSoundManager {
   private audioPool: AudioPool
-  private readonly POOL_SIZE = 3 // Multiple instances for overlapping sounds
+  private readonly POOL_SIZE = 3
   private lastPlayTime = 0
   private readonly DEBOUNCE_MS = 100
   private isInitialized = false
@@ -26,23 +19,17 @@ class ChessSoundManager {
     }
   }
 
-  /**
-   * Initialize audio pool - call this once when app starts
-   */
   public async initialize(): Promise<void> {
     if (this.isInitialized) return
 
     try {
-      // Create audio pools
       for (let i = 0; i < this.POOL_SIZE; i++) {
         const moveAudio = new Audio('/assets/sound/move.mp3')
         const captureAudio = new Audio('/assets/sound/capture.mp3')
 
-        // Preload audio
         moveAudio.preload = 'auto'
         captureAudio.preload = 'auto'
 
-        // Add error handling
         moveAudio.addEventListener('error', (e) =>
           console.warn('Failed to load move sound:', e),
         )
@@ -60,29 +47,23 @@ class ChessSoundManager {
     }
   }
 
-  /**
-   * Play chess move sound with debouncing and capture detection
-   */
   public playMoveSound(isCapture = false): void {
-    // Check if sounds are enabled in user settings
     if (typeof window !== 'undefined') {
       try {
         const settings = localStorage.getItem('maia-user-settings')
         if (settings) {
           const userSettings = JSON.parse(settings)
           if (!userSettings.soundEnabled) {
-            return // Sound disabled in settings
+            return
           }
         }
       } catch (error) {
-        // If settings parsing fails, continue with default behavior
         console.warn('Failed to read sound settings:', error)
       }
     }
 
     const now = Date.now()
 
-    // Debounce rapid sounds
     if (now - this.lastPlayTime < this.DEBOUNCE_MS) {
       return
     }
@@ -100,24 +81,19 @@ class ChessSoundManager {
       return
     }
 
-    // Find available audio instance (not currently playing)
     let audioToPlay = audioPool.find((audio) => audio.paused || audio.ended)
 
-    // If all are playing, use the first one (interrupt)
     if (!audioToPlay) {
       audioToPlay = audioPool[0]
     }
 
     try {
-      // Reset audio to beginning
       audioToPlay.currentTime = 0
 
-      // Play sound
       const playPromise = audioToPlay.play()
 
       if (playPromise !== undefined) {
         playPromise.catch((error) => {
-          // Handle autoplay restrictions (common on mobile)
           if (error.name !== 'AbortError') {
             console.warn(`Failed to play ${soundType} sound:`, error)
           }
@@ -130,9 +106,6 @@ class ChessSoundManager {
     }
   }
 
-  /**
-   * Cleanup resources when app shuts down
-   */
   public cleanup(): void {
     Object.values(this.audioPool)
       .flat()
@@ -146,20 +119,13 @@ class ChessSoundManager {
     this.isInitialized = false
   }
 
-  /**
-   * Check if sound manager is ready
-   */
   public get ready(): boolean {
     return this.isInitialized
   }
 }
 
-// Global singleton instance
 export const chessSoundManager = new ChessSoundManager()
 
-/**
- * React hook for chess sound management
- */
 export const useChessSoundManager = () => {
   return {
     playMoveSound: chessSoundManager.playMoveSound.bind(chessSoundManager),
